Integrated veacy investor api source

+++ RoleRest/src/main/java/com/kaytes/veacy/controller/RoleModuleMappingController.java
@@ -12,14 +12,19 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 import com.kaytes.veacy.dto.ApiReturnResponse;
 import com.kaytes.veacy.dto.RoleModuleMappingApiResponse;
+import com.kaytes.veacy.dto.request.RoleModuleMappingModel;
 import com.kaytes.veacy.entity.RoleModuleMapping;
-import com.kaytes.veacy.model.RoleModuleMappingModel;
 
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+
+/**
+ * The RoleModuleMappingController interface is used for customizing the Swagger
+ */
+
 
 @Tag(name = "RoleModuleMapping Controller" , description = "This Swagger is for RoleModuleMapping Controller")
 public interface RoleModuleMappingController {
@@ -37,21 +42,21 @@ public interface RoleModuleMappingController {
 	@ApiResponses(value = {@ApiResponse(responseCode="200" , description = "Successfully Completed the Task" ,
 			content = {@Content(mediaType = "application/json" ,
 			schema = @Schema(implementation = RoleModuleMapping.class))})})
-	public ResponseEntity<RoleModuleMappingApiResponse> getByRoleId(@PathVariable int id);
+	public ResponseEntity<RoleModuleMappingApiResponse> getByRoleId(@PathVariable Long id);
 	
 	@ApiResponses(value = {@ApiResponse(responseCode="200" , description = "Successfully Completed the Task" ,
 			content = {@Content(mediaType = "application/json" ,
 			schema = @Schema(implementation = RoleModuleMapping.class))})})
-	public ResponseEntity<RoleModuleMappingApiResponse> getBymoduleId(@PathVariable int id);
+	public ResponseEntity<RoleModuleMappingApiResponse> getBymoduleId(@PathVariable Long id);
 		
 	@ApiResponses(value = {@ApiResponse(responseCode="200" , description = "Successfully Completed the Task" ,
 			content = {@Content(mediaType = "application/json" ,
 			schema = @Schema(implementation = RoleModuleMapping.class))})})
-	public ResponseEntity<ApiReturnResponse> deleteRoleMapping(@PathVariable int id);
+	public ResponseEntity<ApiReturnResponse> deleteRoleMapping(@PathVariable Long id);
 	
 	@ApiResponses(value = {@ApiResponse(responseCode="200" , description = "Successfully Completed the Task" ,
 			content = {@Content(mediaType = "application/json" ,
 			schema = @Schema(implementation = RoleModuleMapping.class))})})
-	public ResponseEntity<ApiReturnResponse> update(@PathVariable int id, @RequestBody Map<String, Object> update);
+	public ResponseEntity<ApiReturnResponse> update(@PathVariable Long id, @RequestBody Map<String, Object> update);
 
 }
